requirements.txt
streamlit>=1.37.0
openai>=1.35.0
python-docx>=1.1.0
python-pptx>=0.6.23
